Padronizado a subcategoria

+++ Controllers/SubCategoriaController.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using PetApi.Data;
 using PetApi.Models;
+using PetApi.ViewModels;
 using PetApi.ViewModels.SubcategoriaVM;
 
 namespace PetApi.Controllers;
@@ -24,13 +25,13 @@ public class SubCategoriaController : ControllerBase
         try
         {
             if (_context.SubCategorias == null)
-                return NotFound();
-            
-            return await _context.SubCategorias.ToListAsync();
+                return NotFound(new ResultViewModel<SubCategoria>("Não existe nenhuma subcategoria"));
+            var subcategorias = await _context.SubCategorias.ToListAsync();
+            return Ok(new ResultViewModel<List<SubCategoria>>(subcategorias));
         }
-        catch(Exception e)
+        catch
         {
-            return StatusCode(500, e);
+            return StatusCode(500, new ResultViewModel<SubCategoria>("Falha interna do servidor"));
         }
        
     }
@@ -38,52 +39,63 @@ public class SubCategoriaController : ControllerBase
     [HttpGet("{id}")]
     public async Task<ActionResult<SubCategoria>> GetSubCategoria(int id)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(new ResultViewModel<SubCategoria>("Falha interna no servidor"));
         try
         {
-            var subCategoria = await _context.SubCategorias.FindAsync(id);
-            if (subCategoria == null)
-                return Problem("Subcategoria nula");
+            var subcategoria = await _context.SubCategorias.FindAsync(id);
+            if (subcategoria == null)
+                return NotFound(new ResultViewModel<SubCategoria>("Não existe essa subcategoria"));
 
-            var categoria = await _context.Categorias.FindAsync(subCategoria.CategoriaId);
-            subCategoria.Categoria = categoria;
-            
-            return subCategoria;
+            var categoria = await _context.Categorias.FindAsync(subcategoria.CategoriaId);
+            subcategoria.Categoria = categoria;
+
+            return Ok(new ResultViewModel<SubCategoria>(subcategoria));
         }
         catch
         {
-            return StatusCode(500);
-        }
-       
+            return StatusCode(500, new ResultViewModel<SubCategoria>("Falha interna do servidor"));
+        } 
     }
 
     // POST
     [HttpPost]
     public async Task<ActionResult<SubCategoria>> PostSubCategoria(CreateSubcategoriaVM subcategoriaVM)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(new ResultViewModel<SubCategoria>("Falha interna no servidor"));
+
         if (_context.SubCategorias == null)
             return Problem("Entidade adiciona 'ApplicationDbContext.SubCategoria' é nula.");
+        try
+        {
+            var subcategoria = new SubCategoria
+            {
+                SubCategoriaId = 0,
+                Ativo = 0,
+                Nome = subcategoriaVM.Nome,
+                CategoriaId = subcategoriaVM.CategoriaId,
+                Categoria = await _context.Categorias.FindAsync(subcategoriaVM.CategoriaId)
+            };
 
-        var subcategoria = new SubCategoria{
-            SubCategoriaId = 0,
-            Ativo = 0,
-            Nome = subcategoriaVM.Nome,
-            CategoriaId = subcategoriaVM.CategoriaId,
-            Categoria = await _context.Categorias.FindAsync(subcategoriaVM.CategoriaId)
-        };    
-        
-        if(subcategoria.Categoria == null)
-            return Problem("Categoria não existe");
+            _context.SubCategorias.Add(subcategoria);
+            await _context.SaveChangesAsync();
 
-        _context.SubCategorias.Add(subcategoria);
-        await _context.SaveChangesAsync();
-
-        return CreatedAtAction("GetSubCategoria", new { id = subcategoria.SubCategoriaId }, subcategoria);
+            return CreatedAtAction("GetSubCategoria", new { id = subcategoria.SubCategoriaId }, new ResultViewModel<SubCategoria>(subcategoria));
+        }
+        catch
+        {
+            return StatusCode(500, new ResultViewModel<SubCategoria>("Falha interna do servidor"));
+        }
     }
 
     // PUT:
     [HttpPut("{id}")]
     public async Task<IActionResult> Put(int id, UpdateSubcategoriaVM subcategoriaVM)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(new ResultViewModel<SubCategoria>("Falha interna no servidor"));
+
         var subcategoria = new SubCategoria
         {
             SubCategoriaId = id,
@@ -105,30 +117,37 @@ public class SubCategoriaController : ControllerBase
         catch (DbUpdateConcurrencyException)
         {
             if (!SubCategoriaExists(id))
-                return NotFound("");
-            
-            return BadRequest("01xE5 - Falha ao alterar subcategoria");
-            
+                return NotFound(new ResultViewModel<SubCategoria>("Subcategoria não existe"));
+
+            return BadRequest(new ResultViewModel<Produto>("Falha ao alterar subcategoria"));
         }
 
-        return NoContent();
+        return Ok(new ResultViewModel<SubCategoria>("Subcategoria alterada"));
     }
 
     //Delete
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteSubCategoria(int id)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(new ResultViewModel<SubCategoria>("Falha interna no servidor"));
+
         if (_context.SubCategorias == null)
-            return NotFound();
-        
-        var subcategoria = await _context.SubCategorias.FindAsync(id);
-        if (subcategoria == null)
-            return NotFound();
+            return NotFound(new ResultViewModel<SubCategoria>("Subcategoria nula"));
+        try
+        {
+            var subcategoria = await _context.SubCategorias.FindAsync(id);
+            if (subcategoria == null)
+                return NotFound(new ResultViewModel<SubCategoria>("Subcategoria não encontrada"));
 
-        _context.SubCategorias.Remove(subcategoria);
-        await _context.SaveChangesAsync();
-
-        return NoContent();
+            _context.SubCategorias.Remove(subcategoria);
+            await _context.SaveChangesAsync();
+        }
+        catch
+        {
+            return StatusCode(500, new ResultViewModel<SubCategoria>("Falha interna do servidor"));
+        }
+        return Ok(new ResultViewModel<SubCategoria>("Subcategoria excluída"));
     }
     private bool SubCategoriaExists(int id)
     {
